add func to convert temp
